Birthday Party Service Logic Implementation

+++ src/main/java/edu/icet/repository/EventRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.sql.Date;
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<EventEntity,Integer> {
+public interface EventRepository extends JpaRepository<EventEntity,Long> {
 
     List<EventEntity> findAllByEventType(EventType eventType);
     List<EventEntity> findAllByBudgetType(BudgetType budgetType);
